pyproject.toml
[build-system]
requires = ["setuptools>=69", "setuptools_scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

# Configure isort to be compatible with the "black" code formatter
[tool.isort]
profile = "black"
src_paths = ["src", "tests"]
honor_noqa = true

# Make use of setuptools_scm to dynamically generate the
# package version from git tags.
[tool.setuptools_scm]

[project]
name = "govtech-csg-xcg-modelpermissions"
dynamic = ["version"] # Set the version dynamically based on the most recent git tag (uses setuptools_scm)
description = "Package containing functionality related to django model permissions."
readme = "README.md"
requires-python = ">=3.8" # To align with Django 4.1 and above's requirement
dependencies = [
    "django ~= 4.1",
    "django-crum",
    "django-guardian",
]

classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
]

[project.urls]
Homepage = "https://xcg.tech.gov.sg"
Documentation = "https://xcg.tech.gov.sg"
Repository = "https://github.com/GovTech-CSG/govtech-csg-xcg-modelpermissions"
